refactor: rearrange source files
