+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/injection/CrashModuleImpl.kt
@@ -7,7 +7,6 @@ import io.embrace.android.embracesdk.internal.crash.CrashFileMarkerImpl
 import io.embrace.android.embracesdk.internal.crash.LastRunCrashVerifier
 import io.embrace.android.embracesdk.internal.ndk.NativeCrashDataSourceImpl
 import io.embrace.android.embracesdk.internal.ndk.NativeCrashService
-import io.embrace.android.embracesdk.internal.ndk.NativeModule
 import io.embrace.android.embracesdk.internal.ndk.NoopNativeCrashService
 
 internal class CrashModuleImpl(
