+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/injection/NativeModule.kt
@@ -1,0 +1,13 @@
+package io.embrace.android.embracesdk.internal.injection
+
+import io.embrace.android.embracesdk.internal.anr.ndk.NativeAnrOtelMapper
+import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerInstaller
+import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerService
+import io.embrace.android.embracesdk.internal.ndk.NdkService
+
+internal interface NativeModule {
+    val ndkService: NdkService
+    val nativeThreadSamplerService: NativeThreadSamplerService?
+    val nativeThreadSamplerInstaller: NativeThreadSamplerInstaller?
+    val nativeAnrOtelMapper: NativeAnrOtelMapper
+}
