+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/storage/StorageAvailabilityChecker.kt
@@ -1,20 +1,5 @@
 package io.embrace.android.embracesdk.internal.storage
 
-import android.content.Context
-import android.os.StatFs
-
 internal interface StorageAvailabilityChecker {
     fun getAvailableBytes(): Long
-}
-
-internal class StatFsAvailabilityChecker(
-    context: Context
-) : StorageAvailabilityChecker {
-    private val statFs: StatFs by lazy {
-        StatFs(context.filesDir.path)
-    }
-
-    override fun getAvailableBytes(): Long {
-        return statFs.availableBytes
-    }
 }
